changed the location of function
